# docs/api.yaml
openapi: 3.0.3
info:
  title: OSCAR API
  version: v2.0.0
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0'
  description: OSCAR API documentation
  contact:
    name: Sebas Risco
    email: serisgal@i3m.upv.es
...
paths:
  /system/services:
    get:
      summary: List services
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
      operationId: ListServices
      description: List all created services
      security:
        - basicAuth: []
      tags:
        - services
    post:
      summary: Create service
      operationId: CreateService
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
      description: Create a service
      security:
        - basicAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Service'
      tags:
        - services
    put:
      summary: Update service
      operationId: UpdateService
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      description: Update a service
      security:
        - basicAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Service'
      tags:
        - services
  '/system/services/{serviceName}':
    parameters:
      - schema:
          type: string
        name: serviceName
        in: path
        required: true
    get:
      summary: Read service
      tags:
        - services
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      operationId: ReadService
      security:
        - basicAuth: []
      description: Read a service
    delete:
      summary: Delete service
      operationId: DeleteService
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      description: Delete a service
      security:
        - basicAuth: []
      tags:
        - services
...
components:
  schemas:
    Service:
      title: Service
      type: object
      properties:
        name:
          type: string
        memory:
          type: string
        cpu:
          type: string
        log_level:
          type: string
        image:
          type: string
        script:
          type: string
        environment:
          type: object
          properties:
            Variables:
              type: object
              properties:
                '':
                  allOf:
                    - {}
                  type: string
        input:
          type: array
          items:
            $ref: '#/components/schemas/StorageIOConfig'
        output:
          type: array
          items:
            $ref: '#/components/schemas/StorageIOConfig'
        storage_providers:
          $ref: '#/components/schemas/StorageProviders'
      required:
        - name
        - image
        - script
    StorageIOConfig:
      title: StorageIOConfig
      type: object
      properties:
        storage_provider:
          type: string
        path:
          type: string
        suffix:
          type: array
          items:
            type: string
        prefix:
          type: array
          items:
            type: string
    StorageProviders:
      title: StorageProviders
      type: object
      properties:
        s3:
          type: object
          properties:
            id:
              type: object
              properties:
                access_key:
                  type: string
                secret_key:
                  type: string
                region:
                  type: string
        minio:
          type: object
          properties:
            id:
              $ref: '#/components/schemas/MinIOProvider'
        onedata:
          type: object
          properties:
            id:
              type: object
              properties:
                oneprovider_host:
                  type: string
                token:
                  type: string
                space:
                  type: string
...
    MinIOProvider:
      title: MinIOProvider
      type: object
      properties:
        endpoint:
          type: string
        region:
          type: string
        secret_key:
          type: string
        access_key:
          type: string
        verify:
          type: boolean
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: ''
...
servers:
  - url: 'http://localhost:8080'
    description: ''